Removed ListActivityPrevozov. Now it's in a fragment.
