+++ app/src/main/AndroidManifest.xml
@@ -52,11 +52,6 @@
         </activity>
         <activity android:name=".RegisterActivity" />
         <activity android:name=".MainActivity" />
-        <activity android:name=".ListPrevozovActivity">
-            <meta-data
-                android:name="android.app.searchable"
-                android:resource="@xml/searchable" />
-        </activity>
 
         <receiver android:name=".SmsReceiver">
             <intent-filter android:priority="1">
