+++ app/src/main/res/layout/activity_main.xml
@@ -7,34 +7,6 @@
     tools:context="nameplaceholder.prevazanjaorg.MainActivity">
 
     <Button
-        android:id="@+id/btnIscem"
-        android:layout_width="100dp"
-        android:layout_height="50dp"
-        android:layout_alignBaseline="@+id/btnPonujam"
-        android:layout_alignBottom="@+id/btnPonujam"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:layout_marginLeft="46dp"
-        android:layout_marginStart="46dp"
-        android:onClick="listPrevozovActivity"
-        android:text="IŠČEM"
-        android:textSize="15sp" />
-
-    <Button
-        android:id="@+id/btnPonujam"
-        android:layout_width="100dp"
-        android:layout_height="50dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginBottom="120dp"
-        android:layout_marginEnd="34dp"
-        android:layout_marginRight="34dp"
-        android:onClick="listPrevozovActivity"
-        android:text="PONUJAM"
-        android:textSize="15sp" />
-
-    <Button
         android:id="@+id/btnLogout"
         android:layout_width="100dp"
         android:layout_height="50dp"
@@ -43,7 +15,6 @@
         android:layout_marginStart="23dp"
         android:layout_toEndOf="@+id/btnLanding"
         android:layout_toRightOf="@+id/btnLanding"
-        android:onClick="listPrevozovActivity"
         android:text="LOGOUT"
         android:textSize="15sp" />
 
@@ -52,7 +23,6 @@
         android:layout_width="100dp"
         android:layout_height="50dp"
         android:layout_alignParentBottom="true"
-        android:onClick="listPrevozovActivity"
         android:text="Landing"
         android:textSize="15sp" />
 
@@ -60,9 +30,9 @@
         android:id="@+id/txtUsername"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignLeft="@+id/btnIscem"
+        android:layout_alignLeft="@+id/TestPrijava"
         android:layout_alignParentTop="true"
-        android:layout_alignStart="@+id/btnIscem"
+        android:layout_alignStart="@+id/TestPrijava"
         android:layout_marginTop="20dp"
         android:textSize="24sp" />
 
@@ -88,10 +58,8 @@
         android:id="@+id/TestPrijava"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_above="@+id/btnPonujam"
+        android:layout_above="@+id/btnLanding"
         android:layout_marginBottom="22dp"
-        android:layout_toEndOf="@+id/btnIscem"
-        android:layout_toRightOf="@+id/btnIscem"
         android:text="Test prijava" />
 
     <ToggleButton
