+++ app/src/main/res/values/colors.xml
@@ -11,7 +11,7 @@
     <color name="colorPrimaryDarkish">#e2e2e2</color>
     <color name="colorGrey">#535353</color>
     <color name="colorPrimaryLight">#ffffff</color>
-    <color name="colorAccent">#a6ff4d</color>
+    <color name="colorAccent">#101010</color>
     <color name="colorGreen">#a6ff4d</color>
     <color name="yellow1">#ffce00</color>
     <color name="yellow2">#ffd700</color>
